0wlmusic/requirements.txt
pynacl
discord.py [voice, speed] @ git+https://github.com/Rapptz/discord.py
pip
yt-dlp
urllib3
python-dotenv